api-py/requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-camel-case>=1.4,<2.0
django-cors-headers>=4.3,<5.0
django-filter>=24.3,<25.0
psycopg[binary]>=3.1,<4.0
python-dotenv>=1.0,<2.0
argon2-cffi>=23.1,<24.0
anthropic>=0.40,<1.0
gunicorn>=22.0,<23.0
